pyproject.toml
[tool.poetry]
name = "pymglasso"
version = "0.1.0"
description = "Inference"
authors = ["Edmond"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
scipy = "==1.7.1"
numpy = ">=1.16.5,<1.23.0"
six = "^1.16.0"
pylearn-parsimony = {git = "https://github.com/neurospin/pylearn-parsimony.git"}
scikit-learn = "^1.1.2"
matplotlib = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
